cleaned up AverageStars code
